Update to login.py
Given that the Y1 and Y2 values need to be randomly-generated within the class, I think we may be in a chicken
-egg situation with respect to new_acct(): if a user needs to create a new account, they don't know the Y1 or Y2
values but the __init__ constructor for account requires the Y1 and Y2 values be passed as parameters to the object
constructor. I think we can have it both ways, however; I moved the random_y() method up to a class method of account
and created a from_scratch() class method as an alternate constructor if an account is being created for the first
time. This alternate constructor generates random y values from the class random_y method then passes all of the
correct parameters to the __init__ constructor in the correct order. This way, the sel_acct() function will continue
to work as normal as well.

Let me know if I'm totally off-base here as far as the Y values go.